Hopefully final version

+++ XMega-MS56XX-Driver/src/Drivers/timercounter.h
@@ -12,6 +12,6 @@
 #include <asf.h>
 
 void TC0_setup(TC0_t* TC, enum sysclk_port_id sysclk_port, uint8_t pins_to_ctrl);
-void TC_config(TC0_t* TC, float freq, uint8_t duty);
+void TC_config(TC0_t* TC, float freq, float duty);
 
 #endif /* TIMERCOUNTER_H_ */
